backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
SQLAlchemy==2.0.23
Flask-SocketIO==5.3.6
eventlet==0.36.1
python-socketio==5.10.0
flask-cors==4.0.0
openai>=1.3.0
python-dotenv>=1.0.0
